remove collector and add the client
