--- Packs/NetQuestOMX/Integrations/NetQuestOMXEventCollector/NetQuestOMXEventCollector_description.md
@@ -1,7 +1,0 @@
-**Important:** Contrary to the Druva integration, DruvaEventCollector is supported by Palo Alto Networks.
-
-## Druva Event Collector
-
-The integration is applicable to customers using Realize Ransomware Recovery module with inSync and Phoenix on Druva Public Cloud.
-
-The API Credentials, which is a combination of the Client ID and Secret Key, are equivalent to user name and password. You can access all the Druva APIs and in turn, access the data stored within Druva products. It is strongly recommended not to share the Client ID and Secret Key with unauthorized sources.
